Delete what's_your_name?
We're concerned that either the apostrophe or the question mark in the name of the file itself is causing problems for others to clone the repository with PyCharm/GitHub. The error message on clone is "Unable to create file dc/what's_your_name?: Invalid argument."

Let's hope this clears things up.